conflicts resolved with the latest dev version
